requirements.txt
click==6.7
python-dotenv==0.7.1

certifi==2018.1.18
chardet==3.0.4
idna==2.6
oauthlib==2.0.6
requests==2.18.4
requests-oauthlib==0.8.0
twython==3.6.0
urllib3==1.22
